+++ app/serializers/fast_recipe_serializer.rb
@@ -1,0 +1,6 @@
+class FastRecipeSerializer
+  include FastJsonapi::ObjectSerializer
+
+  set_type :recipe
+  attributes :name, :links, :skill, :profession_id
+end